added Sign Out button, will work on Sign In style
