+++ app/index.tsx
@@ -1,16 +1,18 @@
 // pages/index.tsx
+"use client";
 import { useState, useEffect } from 'react';
 import Link from 'next/link';
-import { getCurrentUser } from 'aws-amplify/auth';
-import { Flex, Menu, Image } from 'antd';
+import { getCurrentUser, signOut } from 'aws-amplify/auth';
+import { Flex, Menu, Image, Button } from 'antd';
 import type { MenuProps } from 'antd';
 import Sider from 'antd/es/layout/Sider';
 import { HomeOutlined, UserOutlined, TeamOutlined, TagOutlined, ReadOutlined, CommentOutlined, SubnodeOutlined, SettingOutlined } from '@ant-design/icons';
 import i18next from './i18n';
+import { useRouter } from "next/navigation";
 
 const Index: React.FC = () => {
   type MenuItem = Required<MenuProps>['items'][number];
-  
+
   function getItem(
     label: React.ReactNode,
     key: React.Key,
@@ -20,20 +22,20 @@ const Index: React.FC = () => {
     options?: { disabled?: boolean },
   ): MenuItem {
     const itemLabel = children || options?.disabled ? label : <Link href={key.toString()}>{label}</Link>;
-    return { 
-      key, 
-      icon, 
-      children, 
-      label: itemLabel, 
-      type, 
-      disabled: options?.disabled 
+    return {
+      key,
+      icon,
+      children,
+      label: itemLabel,
+      type,
+      disabled: options?.disabled
     } as MenuItem;
   }
 
   const items: MenuProps['items'] = [
     getItem(i18next.t('Menu.Home'), '/channel', <HomeOutlined />),
     getItem(i18next.t('Menu.Marketing'), '/marketing', null, [
-      getItem(i18next.t('Menu.Content'), '/content', <ReadOutlined/>, [
+      getItem(i18next.t('Menu.Content'), '/content', <ReadOutlined />, [
         getItem(i18next.t('Menu.Content.Image'), '/content/image', null),
         getItem(i18next.t('Menu.Content.Text'), '/content/text', null),
         getItem(i18next.t('Menu.Content.Video'), '/content/video', null)
@@ -47,9 +49,10 @@ const Index: React.FC = () => {
       getItem(i18next.t('Menu.Tag'), '/tag', <Image src='/asset/icon/coming_soon.svg' height={24} width={24} />, undefined, undefined, { disabled: true }),
     ], 'group'),
     getItem(i18next.t('Menu.Setting'), '/personalization', <SettingOutlined />),
-  ];  
+  ];
 
   const [loginId, setLoginId] = useState(String);
+  const router = useRouter();
   useEffect(() => {
     const fetchLoginId = async () => {
       try {
@@ -78,7 +81,7 @@ const Index: React.FC = () => {
       acc.push(parentPath);
       return acc;
     }, []);
-    
+
     // Remove the last key (current page) to get parent menu keys
     setOpenKeys(parentKeys.slice(0, -1));
   }, []);
@@ -92,7 +95,7 @@ const Index: React.FC = () => {
   };
 
   return (
-    <Sider style={{ height: '100%', position: 'fixed', left: 0 }}>
+    <Sider style={{ height: '100%', position: 'fixed', left: 0, backgroundColor: '#ffffff' }}>
       <Menu
         onClick={onClick}
         selectedKeys={[current]}
@@ -100,8 +103,30 @@ const Index: React.FC = () => {
         onOpenChange={onOpenChange}
         mode='inline'
         items={items}
-        style={{ height: '100%', borderRight: 0 }}
+        style={{ 
+          borderRight: 0,
+          marginBottom: 0,  // Ensure no margin at bottom
+          paddingBottom: 0  // Ensure no padding at bottom
+        }}
       />
+      <div style={{ 
+        position: 'absolute', 
+        bottom: 0, 
+        width: '100%', 
+        padding: '10px', 
+        backgroundColor: '#ffffff' 
+      }}>
+        <Button
+          type="default"
+          style={{ width: '100%' }}
+          onClick={async () => {
+            await signOut();
+            router.push("/login");
+          }}
+        >
+          {i18next.t('Menu.SignOut')}
+        </Button>
+      </div>
     </Sider>
   );
 };
